requirements.txt
gymnasium==0.29.1
stable-baselines3==2.3.2
opencv-python==4.10.0.84
numpy==1.26.4
mss==10.0.0
pydirectinput==1.0.4
pywin32>=307; platform_system=="Windows"
PyYAML==6.0.2
